+++ paper/src/main/kotlin/dev/nyon/magnetic/Main.kt
@@ -1,6 +1,7 @@
 package dev.nyon.magnetic
 
 import dev.nyon.konfig.config.config
+import dev.nyon.magnetic.compat.AuraSkillsCompat
 import dev.nyon.magnetic.compat.McMMOCompat
 import dev.nyon.magnetic.compat.VeinminerCompat
 import dev.nyon.magnetic.config.*
@@ -46,6 +47,7 @@ class Main : JavaPlugin() {
 
         if (Bukkit.getPluginManager().isPluginEnabled("mcMMO")) McMMOCompat.listenForEvents()
         if (Bukkit.getPluginManager().isPluginEnabled("Veinminer")) VeinminerCompat.listenForEvents()
+        if (Bukkit.getPluginManager().isPluginEnabled("AuraSkills")) AuraSkillsCompat.listenForEvents()
     }
 
     private fun moveConfigToNewPath(newPath: Path) {